requirements.txt
lxml==6.0.2
pillow==12.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.1
python-docx==1.2.0
python-pptx==1.0.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
